// src/openapi/models/GrowthTypes.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Baby Tracker
 * Track everything your baby does
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * An enumeration.
 * @export
 * @enum {string}
 */
export enum GrowthTypes {
    HEIGHT = 'HEIGHT',
    HEAD = 'HEAD',
    WEIGHT = 'WEIGHT'
}

export function GrowthTypesFromJSON(json: any): GrowthTypes {
    return GrowthTypesFromJSONTyped(json, false);
}

export function GrowthTypesFromJSONTyped(json: any, ignoreDiscriminator: boolean): GrowthTypes {
    return json as GrowthTypes;
}

export function GrowthTypesToJSON(value?: GrowthTypes | null): any {
    return value as any;
}