// src/openapi/models/Poop.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Baby Tracker
 * Track everything your baby does
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {exists} from '../runtime';

/**
 *
 * @export
 * @interface Poop
 */
export interface Poop {
    /**
     *
     * @type {number}
     * @memberof Poop
     */
    id?: number;
    /**
     *
     * @type {Date}
     * @memberof Poop
     */
    at?: Date;
    /**
     *
     * @type {number}
     * @memberof Poop
     */
    babyId?: number;
}

export function PoopFromJSON(json: any): Poop {
    return PoopFromJSONTyped(json, false);
}

export function PoopFromJSONTyped(json: any, ignoreDiscriminator: boolean): Poop {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'id': !exists(json, 'id') ? undefined : json['id'],
        'at': !exists(json, 'at') ? undefined : (new Date(json['at'])),
        'babyId': !exists(json, 'baby_id') ? undefined : json['baby_id'],
    };
}

export function PoopToJSON(value?: Poop | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'id': value.id,
        'at': value.at === undefined ? undefined : (value.at.toISOString()),
        'baby_id': value.babyId,
    };
}