// src/openapi/models/Sleep.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Baby Tracker
 * Track everything your baby does
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {exists} from '../runtime';

/**
 *
 * @export
 * @interface Sleep
 */
export interface Sleep {
    /**
     *
     * @type {number}
     * @memberof Sleep
     */
    id?: number;
    /**
     *
     * @type {Date}
     * @memberof Sleep
     */
    startAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof Sleep
     */
    endAt?: Date;
    /**
     *
     * @type {number}
     * @memberof Sleep
     */
    babyId?: number;
}

export function SleepFromJSON(json: any): Sleep {
    return SleepFromJSONTyped(json, false);
}

export function SleepFromJSONTyped(json: any, ignoreDiscriminator: boolean): Sleep {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'id': !exists(json, 'id') ? undefined : json['id'],
        'startAt': !exists(json, 'start_at') ? undefined : (new Date(json['start_at'])),
        'endAt': !exists(json, 'end_at') ? undefined : (new Date(json['end_at'])),
        'babyId': !exists(json, 'baby_id') ? undefined : json['baby_id'],
    };
}

export function SleepToJSON(value?: Sleep | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'id': value.id,
        'start_at': value.startAt === undefined ? undefined : (value.startAt.toISOString()),
        'end_at': value.endAt === undefined ? undefined : (value.endAt.toISOString()),
        'baby_id': value.babyId,
    };
}